refactor: split proto codec facade
